@@ -19,53 +45,93 @@
     "storybook:build": "storybook build"
   },
   "dependencies": {
-    "@blocknote/core": "^0.14.1",
-    "@blocknote/mantine": "^0.14.1",
-    "@blocknote/react": "^0.14.1",
+    "@auth/core": "^0.34.2",
+    "@blocknote/core": "^0.16.0",
+    "@blocknote/mantine": "^0.16.0",
+    "@blocknote/react": "^0.16.0",
+    "@blocknote/server-util": "^0.16.0",
+    "@bomdi/codebox": "^2.0.0",
+    "@clerk/clerk-react": "^5.2.9",
+    "@clerk/nextjs": "^5.2.6",
+    "@convex-dev/auth": "^0.0.61",
+    "@editorjs/attaches": "^1.3.0",
+    "@editorjs/checklist": "^1.6.0",
+    "@editorjs/delimiter": "^1.4.2",
+    "@editorjs/editorjs": "^2.30.6",
+    "@editorjs/embed": "^2.7.4",
+    "@editorjs/header": "^2.8.7",
+    "@editorjs/image": "^2.9.3",
+    "@editorjs/inline-code": "^1.5.1",
+    "@editorjs/link": "^2.6.2",
+    "@editorjs/list": "^1.10.0",
+    "@editorjs/quote": "^2.7.2",
+    "@editorjs/simple-image": "^1.6.0",
     "@mantine/code-highlight": "^7.8.1",
-    "@mantine/core": "^7.7.1",
+    "@mantine/core": "^7.11.0",
     "@mantine/dates": "7.4.0",
     "@mantine/dropzone": "^7.4.0",
-    "@mantine/form": "^7.4.0",
-    "@mantine/hooks": "^7.7.1",
+    "@mantine/form": "^7.11.2",
+    "@mantine/hooks": "^7.11.0",
+    "@mantine/modals": "^7.11.1",
     "@mantine/notifications": "^7.4.0",
     "@mdx-js/loader": "^3.0.0",
+    "@million/lint": "^1.0.0-rc.76",
     "@next/mdx": "^14.0.3",
     "@next/third-parties": "^14.0.4",
     "@tabler/icons-react": "^2.45.0",
+    "@types/react-infinite-scroller": "^1.2.5",
+    "@types/swagger-ui-react": "^4.18.3",
     "@vercel/edge-config": "^1.1.0",
     "@vercel/kv": "^1.0.1",
+    "@vercel/speed-insights": "^1.0.12",
+    "chunk-text": "^2.0.1",
+    "clsx": "^2.1.1",
+    "cohere-ai": "^7.14.0",
+    "convex": "1.16.0",
+    "convex-helpers": "^0.1.55",
     "cookies-next": "^2.0.4",
     "cx": "^22.4.1",
     "dayjs": "^1.11.10",
+    "editorjs-blocks-react-renderer": "^1.3.0",
+    "editorjs-html": "^3.4.3",
+    "editorjs-react-renderer": "^3.5.1",
     "firebase": "^9.7.0",
     "firebase-admin": "^12.0.0",
     "jwt-decode": "^3.1.2",
     "million": "^2.6.4",
     "nanoid": "^3.3.4",
-    "next": "14.0.3",
+    "next": "^14.2.11",
     "next-pwa": "^5.6.0",
+    "openai": "^4.67.1",
+    "patch-package": "^8.0.0",
     "posthog-js": "^1.103.1",
-    "prosemirror-model": "1.21.1",
+    "posthog-node": "^4.0.1",
+    "prosemirror-model": "1.22.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
+    "react-editor-js": "1.10.0",
     "react-hook-form": "^7.43.9",
+    "react-infinite-scroller": "^1.2.6",
     "react-masonry-css": "^1.0.16",
     "react-qr-code": "^2.0.12",
+    "redoc": "^2.1.5",
     "sharp": "^0.30.7",
+    "svix": "^1.24.0",
+    "swagger-ui": "^5.17.14",
+    "swagger-ui-react": "^5.17.14",
     "tabler-icons-react": "^1.56.0"
   },
   "devDependencies": {
     "@babel/core": "^7.17.8",
     "@next/bundle-analyzer": "^14.0.3",
     "@next/eslint-plugin-next": "^12.1.4",
-    "@storybook/addon-essentials": "^7.6.7",
-    "@storybook/addon-interactions": "^7.6.7",
-    "@storybook/addon-links": "^7.6.7",
-    "@storybook/addon-onboarding": "^1.0.10",
-    "@storybook/addon-styling-webpack": "^0.0.6",
-    "@storybook/nextjs": "^7.6.7",
-    "@storybook/react": "^7.6.7",
+    "@storybook/addon-essentials": "^8.2.1",
+    "@storybook/addon-interactions": "^8.2.1",
+    "@storybook/addon-links": "^8.2.1",
+    "@storybook/addon-onboarding": "^8.2.1",
+    "@storybook/addon-styling-webpack": "^1.0.0",
+    "@storybook/nextjs": "^8.2.1",
+    "@storybook/react": "^8.2.1",
     "@testing-library/dom": "^8.12.0",
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^13.0.0",
@@ -98,14 +164,18 @@
     "prettier": "^2.6.2",
     "prettier-plugin-tailwindcss": "^0.1.8",
     "react-docgen": "^5.4.1",
-    "storybook": "^7.6.7",
+    "storybook": "^8.2.1",
     "storybook-addon-turbo-build": "^2.0.1",
-    "storybook-dark-mode": "^3.0.3",
+    "storybook-dark-mode": "^4.0.2",
     "tailwindcss": "^3.0.24",
     "ts-jest": "^27.1.4",
-    "typescript": "4.6.3"
+    "typescript": "^5.5.3",
+    "oauth4webapi": "2.12.0"
   },
-  "resolutions": {
-    "**/prosemirror-model": "1.21.1"
+  "overrides": {
+    "oauth4webapi": "2.12.0"
+  },
+  "patchedDependencies": {
+    "oauth4webapi@2.12.0": "patches/oauth4webapi@2.12.0.patch"
   }
 }
