@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const n=(n,a)=>(n=new URL(n+".js",a).href,s[n]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=n,e.onload=s,document.head.appendChild(e)}else e=n,importScripts(n),s()})).then((()=>{let e=s[n];if(!e)throw new Error(`Module ${n} didn’t register its module`);return e})));self.define=(a,c)=>{const i=e||("document"in self?document.currentScript.src:"")||location.href;if(s[i])return;let r={};const t=e=>n(e,i),d={module:{uri:i},exports:r,require:t};s[i]=Promise.all(a.map((e=>d[e]||t(e)))).then((e=>(c(...e),r)))}}define(["./workbox-2780d724"],(function(e){"use strict";importScripts("fallback-ARt2Ft0JFck4D99NzF8OE.js"),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/static/ARt2Ft0JFck4D99NzF8OE/_buildManifest.js",revision:"dba9a49881cd7ade4ef698b4986da3e1"},{url:"/_next/static/ARt2Ft0JFck4D99NzF8OE/_middlewareManifest.js",revision:"fb2823d66b3e778e04a3f681d0d2fb19"},{url:"/_next/static/ARt2Ft0JFck4D99NzF8OE/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/125-01f967f93af17d58.js",revision:"01f967f93af17d58"},{url:"/_next/static/chunks/185-1a0855067d88208e.js",revision:"1a0855067d88208e"},{url:"/_next/static/chunks/19.2d2354e5a7e7eb62.js",revision:"2d2354e5a7e7eb62"},{url:"/_next/static/chunks/336-68d882e8b2c89d93.js",revision:"68d882e8b2c89d93"},{url:"/_next/static/chunks/361-0290b3defbcc4a81.js",revision:"0290b3defbcc4a81"},{url:"/_next/static/chunks/397-f02acbbbca6937d1.js",revision:"f02acbbbca6937d1"},{url:"/_next/static/chunks/506-a6561bbcfd191180.js",revision:"a6561bbcfd191180"},{url:"/_next/static/chunks/729-b0acd9441b23a8cc.js",revision:"b0acd9441b23a8cc"},{url:"/_next/static/chunks/808-f11ed5f529c9f503.js",revision:"f11ed5f529c9f503"},{url:"/_next/static/chunks/863-58cfe5198d5a3c40.js",revision:"58cfe5198d5a3c40"},{url:"/_next/static/chunks/framework-81da43a8dcd978d9.js",revision:"81da43a8dcd978d9"},{url:"/_next/static/chunks/main-b46b69e75fb807c0.js",revision:"b46b69e75fb807c0"},{url:"/_next/static/chunks/pages/%5BauthorId%5D-da8c8669e719a9a4.js",revision:"da8c8669e719a9a4"},{url:"/_next/static/chunks/pages/%5BauthorId%5D/%5BpostId%5D-7ae125858b9ed080.js",revision:"7ae125858b9ed080"},{url:"/_next/static/chunks/pages/404-e6c3f93003842fb8.js",revision:"e6c3f93003842fb8"},{url:"/_next/static/chunks/pages/_error-c0f73fcc0f959fcd.js",revision:"c0f73fcc0f959fcd"},{url:"/_next/static/chunks/pages/_offline-73c17ad55444be8c.js",revision:"73c17ad55444be8c"},{url:"/_next/static/chunks/pages/configuracion-e0080822806f26e2.js",revision:"e0080822806f26e2"},{url:"/_next/static/chunks/pages/crear-472495999055b194.js",revision:"472495999055b194"},{url:"/_next/static/chunks/pages/index-6c09255052c1d26c.js",revision:"6c09255052c1d26c"},{url:"/_next/static/chunks/pages/privacidad-34a1f20001059313.js",revision:"34a1f20001059313"},{url:"/_next/static/chunks/pages/welcome-f9092fa2691bc5bb.js",revision:"f9092fa2691bc5bb"},{url:"/_next/static/chunks/polyfills-5cd94c89d3acac5f.js",revision:"99442aec5788bccac9b2f0ead2afdd6b"},{url:"/_next/static/chunks/webpack-f8aeca60fb35c448.js",revision:"f8aeca60fb35c448"},{url:"/_next/static/css/1158a562700bc7ce.css",revision:"1158a562700bc7ce"},{url:"/_next/static/css/b8e271689cb667ed.css",revision:"b8e271689cb667ed"},{url:"/_offline",revision:"ARt2Ft0JFck4D99NzF8OE"},{url:"/android-chrome-192x192.png",revision:"fe9bd64617ee671bd032edfdd35365f8"},{url:"/android-chrome-512x512.png",revision:"5cdfb78904ff19232d22210cf2cb31e8"},{url:"/apple-touch-icon.png",revision:"79c8ba675e458b745ce6e177ebba487b"},{url:"/favicon-16x16.png",revision:"bfce1cc812f5886ad4ef1012152cff91"},{url:"/favicon-32x32.png",revision:"a18c2a669d849efdb700b4da03204fa2"},{url:"/favicon.ico",revision:"06efcc14ef0900e862daccdada0d91b1"},{url:"/girl.jpg",revision:"9c6ab3a75083f158877256cd0d735988"},{url:"/landscape404.svg",revision:"2f9652f7eaf5a89037580ea3c07f6887"},{url:"/logo.png",revision:"fe9bd64617ee671bd032edfdd35365f8"},{url:"/logologo.svg",revision:"129f56b8730f4748f7ff3b9e7942487a"},{url:"/manifest.json",revision:"5d40d7d82ae2169e121a0403b543c036"},{url:"/perfil.jpg",revision:"b09fc43cd92255a09a1bc579b9075191"},{url:"/porsche.jpg",revision:"debfd154f244e06be10ee1ff841919f2"},{url:"/site.webmanifest",revision:"053100cb84a50d2ae7f5492f7dd7f25e"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:n,state:a})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s},{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600}),{handlerDidError:async({request:e})=>self.fallback(e)}]}),"GET")}));
+if(!self.define){let s,e={};const n=(n,i)=>(n=new URL(n+".js",i).href,e[n]||new Promise((e=>{if("document"in self){const s=document.createElement("script");s.src=n,s.onload=e,document.head.appendChild(s)}else s=n,importScripts(n),e()})).then((()=>{let s=e[n];if(!s)throw new Error(`Module ${n} didn’t register its module`);return s})));self.define=(i,c)=>{const a=s||("document"in self?document.currentScript.src:"")||location.href;if(e[a])return;let t={};const r=s=>n(s,a),u={module:{uri:a},exports:t,require:r};e[a]=Promise.all(i.map((s=>u[s]||r(s)))).then((s=>(c(...s),t)))}}define(["./workbox-1bb06f5e"],(function(s){"use strict";importScripts(),self.skipWaiting(),s.clientsClaim(),s.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"f7900c4a30899b70aacb5d5f673bfc5d"},{url:"/_next/static/9lqufrAsQBiqXY5_n9Tsq/_buildManifest.js",revision:"ff6fd5bbd989a6b8c17a3b8fd89a0b66"},{url:"/_next/static/9lqufrAsQBiqXY5_n9Tsq/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/187-6be20c999103c416.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/2020-743b5bd98fcb2a1c.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/231-f80bef61ae334f87.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/243dde97-04c832c9b6508941.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/3031-13b636cad9c8c153.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/3073-7c4a1acc88073434.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/3103-bf1f91ddd3a920c8.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/3112-bbb08ac113437cfb.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/355-b28c8324fabf4ebc.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/3568-7ea5188220519cd9.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/3904-b368e8d066d4636f.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/433-7553a0e127cd769f.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/4444-b77410290e8a0140.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/4647.6ccf4ee37a1c7fe9.js",revision:"6ccf4ee37a1c7fe9"},{url:"/_next/static/chunks/5423-980f7a05c643e82e.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/5913-74c3a14bba0be46d.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/6468630d-676bc85a008c2cde.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/6591-8773f1bdf6690f7b.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/7023-509bd684d45d4900.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/7143-9b3ea811315cfec0.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/7497-a118c323b0c1bed9.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/766-09ca0645f653c4b1.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/7755-375ba4dbba67e426.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/7915-0cc6716d3900997d.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/798901a7-bc326b8f36f673c4.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/8174-9afba165f462e5c8.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/8283-f6cbdd9447d85de7.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/8371-9a0b2b53f9e96314.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/8903-c455f1bfe258bcb0.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/9107-8ea4d25609e44e4e.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/985-d9f5be5907cf4a28.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/(main)/%5Bauthor%5D/%5BpostSlug%5D/loading-25bb8eeff949f36e.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/(main)/%5Bauthor%5D/%5BpostSlug%5D/page-067c0fc4c034a1dc.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/(main)/%5Bauthor%5D/page-e235bddb4f1db70c.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/(main)/configuracion/page-d83f563c05822153.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/(main)/crear/page-aeb19d5d49662abc.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/(main)/layout-dbff4ef81b50e1a1.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/(main)/page-82c8af802bbe616f.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/_not-found/page-c71624c119945279.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/admin/redireccionar/page-eeca9c29f163e091.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/bienvenido/page-4615301cd2955c30.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/layout-057479979cd47499.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/r/%5Bid%5D/page-7d1f57bfa316fed4.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/socialuapi/page-62a8ff9bf8a629cb.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/app/sso-callback/page-43587b1aa7efbd2d.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/c16f53c3-09754c255da5a582.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/fd9d1056-34dab267c24b7f47.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/framework-08aa667e5202eed8.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/main-28173c1da6717140.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/main-app-2cb5af51f2a2843f.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/pages/_app-f870474a17b7f2fd.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/pages/_error-c66a4e8afc46f17b.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-d382c387bf22e3dc.js",revision:"9lqufrAsQBiqXY5_n9Tsq"},{url:"/_next/static/css/32be159857a3337a.css",revision:"32be159857a3337a"},{url:"/_next/static/css/38213ac1b88a21ea.css",revision:"38213ac1b88a21ea"},{url:"/_next/static/css/3dee52ace510838f.css",revision:"3dee52ace510838f"},{url:"/_next/static/css/54cf597dcfa2c536.css",revision:"54cf597dcfa2c536"},{url:"/_next/static/css/59c11d7fa7556953.css",revision:"59c11d7fa7556953"},{url:"/_next/static/css/8df49318f1de494d.css",revision:"8df49318f1de494d"},{url:"/_next/static/css/a5cd56fcad490037.css",revision:"a5cd56fcad490037"},{url:"/_next/static/css/e82236126d6b4b7b.css",revision:"e82236126d6b4b7b"},{url:"/_next/static/css/eeedbfa49c5cd77c.css",revision:"eeedbfa49c5cd77c"},{url:"/android-chrome-192x192.png",revision:"8c1a4ebe0c1e3cae6d185e3b667401ea"},{url:"/android-chrome-256x256.png",revision:"be854324bc83ad2421dca1d96bcc81a1"},{url:"/android-chrome-512x512.png",revision:"5cdfb78904ff19232d22210cf2cb31e8"},{url:"/apple-touch-icon-precomposed.png",revision:"df9318e1707a1eb62a6da72a2b451ef1"},{url:"/apple-touch-icon.png",revision:"0ab5a2250b08fe87b6c503c81a1066d4"},{url:"/browserconfig.xml",revision:"61bfd064535af0c276bb63b3fd579733"},{url:"/convex-spec-1725562104988.yaml",revision:"8f24be0a3d66f59d3b62897aefe95740"},{url:"/favicon-16x16.png",revision:"deedeb40d989553507d03c23523601ee"},{url:"/favicon-32x32.png",revision:"7a2154b491939ca6850a6b28b2c3bad7"},{url:"/favicon.ico",revision:"2e3f93813bf157f6335eb05284307800"},{url:"/girl.jpg",revision:"9c6ab3a75083f158877256cd0d735988"},{url:"/html_code.html",revision:"3a16e722a77b37f97c3564c145f55497"},{url:"/landscape404.svg",revision:"2f9652f7eaf5a89037580ea3c07f6887"},{url:"/logo.png",revision:"fe9bd64617ee671bd032edfdd35365f8"},{url:"/logologo.svg",revision:"b61f21fff624be7e78ec2d72dd14efae"},{url:"/manifest.json",revision:"46ad466319bfb2d3c0a574c9687eec4a"},{url:"/mstile-150x150.png",revision:"a17e09848072258afb448acfe037ea63"},{url:"/perfil.jpg",revision:"b09fc43cd92255a09a1bc579b9075191"},{url:"/porsche.jpg",revision:"debfd154f244e06be10ee1ff841919f2"},{url:"/reco.jpg",revision:"30073a6f3534f6cf6b9c65e2658191a8"},{url:"/safari-pinned-tab.svg",revision:"fd112cdbb76650768d4ce1856abddb46"},{url:"/site.webmanifest",revision:"2b80778dade1bccefe3d4057686b14a7"}],{ignoreURLParametersMatching:[]}),s.cleanupOutdatedCaches(),s.registerRoute("/",new s.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:s,response:e,event:n,state:i})=>e&&"opaqueredirect"===e.type?new Response(e.body,{status:200,statusText:"OK",headers:e.headers}):e}]}),"GET"),s.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new s.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new s.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),s.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new s.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new s.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),s.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new s.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new s.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),s.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new s.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new s.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),s.registerRoute(/\/_next\/image\?url=.+$/i,new s.StaleWhileRevalidate({cacheName:"next-image",plugins:[new s.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),s.registerRoute(/\.(?:mp3|wav|ogg)$/i,new s.CacheFirst({cacheName:"static-audio-assets",plugins:[new s.RangeRequestsPlugin,new s.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),s.registerRoute(/\.(?:mp4)$/i,new s.CacheFirst({cacheName:"static-video-assets",plugins:[new s.RangeRequestsPlugin,new s.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),s.registerRoute(/\.(?:js)$/i,new s.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new s.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),s.registerRoute(/\.(?:css|less)$/i,new s.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new s.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),s.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new s.StaleWhileRevalidate({cacheName:"next-data",plugins:[new s.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),s.registerRoute(/\.(?:json|xml|csv)$/i,new s.NetworkFirst({cacheName:"static-data-assets",plugins:[new s.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),s.registerRoute((({url:s})=>{if(!(self.origin===s.origin))return!1;const e=s.pathname;return!e.startsWith("/api/auth/")&&!!e.startsWith("/api/")}),new s.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new s.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),s.registerRoute((({url:s})=>{if(!(self.origin===s.origin))return!1;return!s.pathname.startsWith("/api/")}),new s.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new s.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),s.registerRoute((({url:s})=>!(self.origin===s.origin)),new s.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new s.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
