@@ -1,31 +1,36 @@
+"use client"
 import Link from "next/link";
-import { useState } from 'react';
+// import { useState } from 'react';
 import { Group, Code, NavLink } from '@mantine/core';
 import {
   IconLogout,
 } from '@tabler/icons-react';
 // import { MantineLogo } from '@mantinex/mantine-logo';
 import classes from './NavbarSimple.module.css';
-import { useAuth } from "../../context/AuthContext";
-import Protected from "../Protected";
-import Config from "../../config"
+// import { useAuth } from "@context/AuthContext";
+// import Protected from "../Protected";
+import Config from "@lib/config"
+import { Authenticated } from "convex/react";
+import { useClerk } from "@clerk/nextjs";
+import Protected from "@components/Protected";
 
   const data = Config().sidebar
   const version = Config().version
 
 export function AppNavbar() {
   // const { classes, cx } = useStyles();
-  const { logout } = useAuth();
+  const { signOut } = useClerk();
 
 
   const links = data.map((item) => (
-    <NavLink
-        label={item.label}
-        leftSection={<item.icon className={classes.linkIcon} stroke="1.5" />}
-        href={item.link}
-        key={item.label}
-        component={Link}
-      />
+    <Protected.Component on={item.isProtected} key={item.label}>
+      <NavLink
+          label={item.label}
+          leftSection={<item.icon className={classes.linkIcon} stroke="1.5" />}
+          href={item.link}
+          component={Link}
+        />
+    </Protected.Component>
   ));
 
   return (
@@ -45,12 +50,12 @@ export function AppNavbar() {
           <IconSwitchHorizontal className={classes.linkIcon} stroke={1.5} />
           <span>Change account</span>
         </a> */}
-        <Protected.Component>
-          <a href="#" className={classes.link} onClick={() => logout()}>
+        <Authenticated>
+          <a href="#" className={classes.link} onClick={() => signOut()}>
             <IconLogout className={classes.linkIcon} stroke={1.5} />
             <span>Cerrar Sesión</span>
           </a>
-        </Protected.Component>
+        </Authenticated>
       </div>
     </nav>
   );
