@@ -1,61 +1,57 @@
 "use client"
-// import { RichTextEditor, Link } from '@mantine/tiptap';
-// import { useEditor, BubbleMenu, FloatingMenu, Node } from '@tiptap/react';
-// import Highlight from '@tiptap/extension-highlight';
-// import StarterKit from '@tiptap/starter-kit';
-// import Placeholder from '@tiptap/extension-placeholder';
-import { uploadImage, checkImageAdultnessWithMicrosoft } from "../hooks/image"
+import { uploadFileToConvex } from "@hooks/image"
 
+import { Skeleton, useMantineColorScheme } from "@mantine/core";
 import { es } from "./es"
+
+// Add the missing 'emoji' property to the 'slash_menu' dictionary
+// es.slash_menu.emoji = {
+//   title: "Emoji",
+//   subtext: "Insert an emoji",
+//   aliases: ["emoji"],
+//   group: "Insert",
+// };
 import { useCreateBlockNote } from "@blocknote/react";
 import { BlockNoteView } from "@blocknote/mantine";
+import "@blocknote/mantine/style.css";
 
-import { useColorScheme } from "@mantine/hooks";
-import { Skeleton, TypographyStylesProvider, useMantineColorScheme } from "@mantine/core";
-import { useController } from "react-hook-form";
-import { MAXIMUM_MESSAGE_LENGTH, MINIMUM_MESSAGE_LENGTH } from "../constants";
-import { FileDatabase } from "tabler-icons-react";
-// import '@mantine/tiptap/styles.css';
+import { set, useController } from "react-hook-form";
+import { MAXIMUM_MESSAGE_LENGTH, MINIMUM_MESSAGE_LENGTH } from "@lib/constants";
 
-
-// import { useController } from 'react-hook-form';
-// import { MAXIMUM_MESSAGE_LENGTH, MINIMUM_MESSAGE_LENGTH } from '../constants';
-// import { useEffect } from 'react';
-
-
-
-// const content =
-//   '<h2 style="text-align: center;">Welcome to Mantine rich text editor</h2><p><code>RichTextEditor</code> component focuses on usability and is designed to be as simple as possible to bring a familiar editing experience to regular users. <code>RichTextEditor</code> is based on <a href="https://tiptap.dev/" rel="noopener noreferrer" target="_blank">Tiptap.dev</a> and supports all of its features:</p><ul><li>General text formatting: <strong>bold</strong>, <em>italic</em>, <u>underline</u>, <s>strike-through</s> </li><li>Headings (h1-h6)</li><li>Sub and super scripts (<sup>&lt;sup /&gt;</sup> and <sub>&lt;sub /&gt;</sub> tags)</li><li>Ordered and bullet lists</li><li>Text align&nbsp;</li><li>And all <a href="https://tiptap.dev/extensions" target="_blank" rel="noopener noreferrer">other extensions</a></li></ul>';
-
-interface TextEditorProps {
+export interface TextEditorProps {
   name: string;
   required?: boolean;
   editable?: boolean;
-  control: any
+  control: any;
+  setEditorState: React.Dispatch<React.SetStateAction<{ html: string; markdown: string, blocks: any[] } | undefined>>;
 }
 
-const TextEditor = ({ editable, name, required, control }: TextEditorProps) => {
+const editorConfig = {
+  defaultStyles: true,
+  dictionary: es,
+  enableBlockNoteExtensions: true,
+};
+
+const TextEditor = ({ editable, name, control, setEditorState }: TextEditorProps) => {
   const {
     field,
-    formState: { errors },
-    fieldState: { invalid, error },
+    formState,
+    fieldState: { error },
   } = useController({
     name,
     control,
     rules: {
       required: "La descripcion es necesaria",
-      minLength: { value: MINIMUM_MESSAGE_LENGTH * 2, message: `Muy poco contenido, metele más` },
-      maxLength: { value: MAXIMUM_MESSAGE_LENGTH ** 20, message: `Nojoda te pasastes.` }
+      // minLength: { value: MINIMUM_MESSAGE_LENGTH * 2, message: `Muy poco contenido, metele más` },
+      // maxLength: { value: MAXIMUM_MESSAGE_LENGTH ** 20, message: `Nojoda te pasastes.` }
     },
   });
 
   const { colorScheme } = useMantineColorScheme()
 
   const theEditor = useCreateBlockNote({
-    defaultStyles: true,
-    dictionary: es,
-    enableBlockNoteExtensions: true,
-    uploadFile: uploadImage,
+    ...editorConfig,
+    uploadFile: uploadFileToConvex,
   });
 
   const onDocumentChange = async () => {
@@ -64,29 +60,57 @@ const TextEditor = ({ editable, name, required, control }: TextEditorProps) => {
     field.onChange(html);
   };
 
+  const onFormSubmit = async () => {
+    // Converts the editor's contents from Block objects to HTML and store to state.
+    const blocks = theEditor.document;
+    // console.log(blocks);
+    const html = await theEditor.blocksToHTMLLossy(theEditor.document);
+    const markdown = await theEditor.blocksToMarkdownLossy(theEditor.document);
+    setEditorState({ html, markdown, blocks });
+    field.onChange(blocks);
+  }
+
+  const onEditorChange = async () => {
+    field.onChange(theEditor);
+  }
+
+
 
   return (
-    <div className="min-h-32">
-
+    <div className="min-h-32" data-posthog-capture="no">
       <BlockNoteView
         editor={theEditor}
         editable={editable || true}
-        onChange={onDocumentChange}
-        theme={colorScheme}
+        onChange={onEditorChange}
+        theme={colorScheme === "dark" ? "dark" : "light"}
       />
-
       {(error) && <span className="text-red-500">{error.message}</span>}
     </div>
   );
+}
+
+export const ViewPost = ({ content }: { content: any[] }) => {
+  const editor = useCreateBlockNote({
+    ...editorConfig,
+    initialContent: content
+  });
+  const { colorScheme } = useMantineColorScheme()
+  return (
+    <BlockNoteView
+      editor={editor}
+      editable={false}
+      theme={colorScheme === "dark" ? "dark" : "light"}
+    />
+  )
 }
 
 export const EditorLoader = () => {
   return (
     <div className="flex flex-col w-full min-h-32 gap-y-2">
       <Skeleton height="18" width="96%" radius="sm" />
-      <Skeleton height="18" width="92%"radius="sm" />
-      <Skeleton height="18" width="98%"radius="sm" />
-      <Skeleton height="18" width="42%"radius="sm" />
+      <Skeleton height="18" width="92%" radius="sm" />
+      <Skeleton height="18" width="98%" radius="sm" />
+      <Skeleton height="18" width="42%" radius="sm" />
     </div>
   )
 }
